183 refactor maincontentcomponent (#203)
* feat: made organizationCard component

* feat: renamed folder

* feat: moved organizationlist to dashboardpage

* feat: made contactstable component

* feat: Added contactFieldTable

* feat: changed component name

* feat: added ERRroles table component

* feat: added role details component

* feat: rename maincontenctcompoennt to detailedOrgVIew

* feat: Added components folder to dashboard

* feat: Made style files for components

* feat: Changed small import

* feat: moved functions from orgview to relevant component

* feat: Changed component for ContactsSearch

* chore: removed unused functions

* feat: small changes to styling

* fix: small fix

---------
