+++ frontend/altinn-support-dashboard.client/package.json
@@ -21,6 +21,7 @@
     "@mui/icons-material": "^6.1.7",
     "@mui/material": "^6.1.7",
     "@navikt/aksel-icons": "^6.17.0",
+    "@tanstack/react-query": "^5.89.0",
     "mui": "^0.0.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
