+++ frontend/altinn-support-dashboard.client/src/components/Dashboard/styles/ContactsSearchBar.styles.ts
@@ -1,0 +1,16 @@
+import { SxProps } from "@mui/material";
+
+export const searchBarContainerStyle: React.CSSProperties = {
+  display: "flex",
+  alignItems: "center",
+  gap: "8px",
+};
+
+export const textFieldStyle: SxProps = {
+  mb: 2,
+};
+
+export const clearButtonStyle: SxProps = {
+  height: "fit-content",
+  mb: 2,
+};
