+++ frontend/altinn-support-dashboard.client/src/components/Dashboard/utils/contactUtils.ts
